+++ vkconfig/configurator.cpp
@@ -47,11 +47,7 @@ Configurator &Configurator::Get(const std::string &VULKAN_SDK) {
 }
 
 Configurator::Configurator(const std::string &VULKAN_SDK)
-    : path(VULKAN_SDK, SUPPORTED_CONFIG_FILES),
-      environment(path),
-      layers(environment),
-      configurations(environment),
-      request_vulkan_status(true) {}
+    : path(VULKAN_SDK, SUPPORTED_CONFIG_FILES), environment(path), layers(environment), configurations(environment) {}
 
 Configurator::~Configurator() {
     configurations.SaveAllConfigurations(layers.available_layers);
@@ -127,8 +123,6 @@ void Configurator::ActivateConfiguration(const std::string &configuration_name) 
             ActivateConfiguration(configuration->key);
         }
     }
-
-    this->request_vulkan_status = true;
 }
 
 bool Configurator::SupportDifferentLayerVersions(Version *return_loader_version) const {