+++ vkconfig/CHANGELOG.md
@@ -15,6 +15,8 @@
 
 ### Improvements:
 - Add 'status' to give an explicit control to print the Vulkan status log
+- Add support for command line outputs on Windows when using 'Command Prompt'
+- Add support for using Vulkan Configurator environment variables in the UI
 
 ### Fixes:
 - Fix clear log when changing `Vulkan Kiader Messages` type when `Clear log on action` is unchecked